+++ src/ApiCall.js
@@ -14,7 +14,7 @@ const api = new Proxy(
     get(_, fnName) {
       return (...params) => {
         const { method, funcName } = parse(fnName);
-        return functionCall.callFn(method, funcName, ...params);
+        return functionCall.callFnResult(method, funcName, ...params);
       };
     },
   }